+++ lib/src/utils/adaptive.dart
@@ -1,7 +1,9 @@
 // Copyright (c) 2025. Alexandr Moroz
 
 import 'dart:math';
+
 import 'package:flutter/cupertino.dart';
+
 import '../theme/constants.dart';
 
 /// Получить размер экрана
@@ -9,8 +11,7 @@ Size screenSize(BuildContext context) => MediaQuery.sizeOf(context);
 
 /// Проверить, является ли экран маленьким в landscape режиме
 bool _smallLandscape(BuildContext context) =>
-    screenSize(context).height < constants.scrXSHeight && 
-    MediaQuery.orientationOf(context) == Orientation.landscape;
+    screenSize(context).height < constants.scrXSHeight && MediaQuery.orientationOf(context) == Orientation.landscape;
 
 /// Проверить, является ли экран большим
 bool isBigScreen(BuildContext context) {
@@ -19,49 +20,22 @@ bool isBigScreen(BuildContext context) {
 }
 
 /// Проверить, можно ли показывать вертикальные панели
-bool canShowVerticalBars(BuildContext context) => 
-    isBigScreen(context) || _smallLandscape(context);
+bool canShowVerticalBars(BuildContext context) => isBigScreen(context) || _smallLandscape(context);
 
 /// Размеры для адаптивности
 enum AdaptiveSize { xxs, xs, s, m, l }
 
 /// Адаптивный контейнер с ограничением ширины
 class MTAdaptive extends StatelessWidget {
-  const MTAdaptive({
-    super.key, 
-    required this.child, 
-    this.force = false, 
-    this.size = AdaptiveSize.m, 
-    this.padding
-  });
+  const MTAdaptive({super.key, required this.child, this.force = false, this.size = AdaptiveSize.m, this.padding});
 
-  const MTAdaptive.xxs({
-    super.key, 
-    required this.child, 
-    this.force = true, 
-    this.padding
-  }) : size = AdaptiveSize.xxs;
+  const MTAdaptive.xxs({super.key, required this.child, this.force = true, this.padding}) : size = AdaptiveSize.xxs;
 
-  const MTAdaptive.xs({
-    super.key, 
-    required this.child, 
-    this.force = true, 
-    this.padding
-  }) : size = AdaptiveSize.xs;
+  const MTAdaptive.xs({super.key, required this.child, this.force = true, this.padding}) : size = AdaptiveSize.xs;
 
-  const MTAdaptive.s({
-    super.key, 
-    required this.child, 
-    this.force = false, 
-    this.padding
-  }) : size = AdaptiveSize.s;
+  const MTAdaptive.s({super.key, required this.child, this.force = false, this.padding}) : size = AdaptiveSize.s;
 
-  const MTAdaptive.l({
-    super.key, 
-    required this.child, 
-    this.force = false, 
-    this.padding
-  }) : size = AdaptiveSize.l;
+  const MTAdaptive.l({super.key, required this.child, this.force = false, this.padding}) : size = AdaptiveSize.l;
 
   final Widget? child;
   final bool force;